requirements/base.txt
Django==1.8.4
django-taggit==0.17.0
django-taggit-serializer==0.1.1
djangorestframework==3.2.3
MySQL-python==1.2.5
